+++ database/migrations/2026_05_08_025526_create_trek_ratings_and_scores_table.php
@@ -25,19 +25,13 @@ return new class extends Migration
         if (!Schema::hasTable('trek_ratings')) {
             Schema::create('trek_ratings', function (Blueprint $table) {
                 $table->id();
-
-                $table->foreignId('tour_package_id')
-                      ->constrained()
-                      ->onDelete('cascade');
-                $table->foreignId('user_id')
-                      ->constrained()
-                      ->onDelete('cascade');
-                $table->foreignId('tour_booking_id')
-                      ->constrained()
-                      ->onDelete('cascade');
-                $table->unsignedTinyInteger('rating');
+                $table->foreignId('tour_package_id')->constrained()->onDelete('cascade');
+                $table->foreignId('user_id')->constrained()->onDelete('cascade');
+                $table->foreignId('tour_booking_id')->constrained()->onDelete('cascade');
+                $table->smallInteger('rating'); 
                 $table->text('review')->nullable();
                 $table->timestamps();
+
                 $table->unique(['tour_package_id', 'user_id'], 'unique_trek_rating');
                 $table->index(['tour_package_id', 'rating'], 'idx_package_rating');
                 $table->index('user_id', 'idx_user');
@@ -55,6 +49,7 @@ return new class extends Migration
                 $table->decimal('final_score',      5, 4)->default(0);
                 $table->timestamp('calculated_at')->nullable();
                 $table->timestamps();
+
                 $table->unique(['user_id', 'tour_package_id']);
                 $table->index(['user_id', 'final_score'], 'idx_user_score');
             });
@@ -68,8 +63,8 @@ return new class extends Migration
 
         if (Schema::hasTable('posts')) {
             Schema::table('posts', function (Blueprint $table) {
-                $table->decimal('rating_avg',   3, 2)->default(0.00);
-                $table->unsignedInteger('rating_count')->default(0);
+                $table->decimal('rating_avg', 3, 2)->default(0.00);
+                $table->integer('rating_count')->default(0); 
             });
         }
     }
